Move PreviewToolbarOverlayTabs into separate file
